requirements.txt
fastapi[all]
nltk
PyPDF2
python-docx